requirements.txt
google-adk[extensions]>=1.24.1
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=5.0.0
pdfplumber>=0.11.0
pymupdf>=1.26.0
pytest>=8.0.0
